% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benef_sim.R
\name{benef_sim}
\alias{benef_sim}
\title{Generate the position of benefitial allele and selection strength.}
\usage{
benef_sim(
  haplotype,
  n_benef = 1,
  min = 0.05,
  max = 0.05,
  fix_sel = 0,
  repli = 1,
  diff_sel_str = FALSE
)
}
\arguments{
\item{haplotype}{Binary numeric matrix containing information of haplotype structure. The columns corresponds to the haplotypes and the rows to the SNPs.}

\item{n_benef}{Non-negative integer, containing information of number of benefitial alleles. 0 corresponds to the scenario of no benifitial alleles.}

\item{min}{Numeric, minimum selection strength of the benefitial allele, must be between 0 and max.}

\item{max}{Numeric, maximum selection strength of the benefitial allele, must be between min and 1.}

\item{fix_sel}{Numeric or numeric vector, whether to fix the number of selected haplotypes to some number. 0 means no fixing, any integer greater than 0 fixes the number of selected haplotypes to the number, any integer lower than 0 means no such number of selected haplotypes (i.e. -1 means no cases of 1 selected haplotypes). Can take numeric vector input, for example c(1,2) will fix the number of selected haplotypes to 1 and 2, c(-1,-2) will disallow number of selected haplotypes as 1 and 2.}

\item{repli}{Numeric, number of replicated populations.}

\item{diff_sel_str}{Logical statement, given there are replicates, whether to allow difference in selection strength between replicates.}
}
\value{
A list, with the first item being a numeric vector of selective strength for all SNPs, the second item being the position of benefitial alleles, the third item being the number of selected haplotypes.
}
\description{
Generate the position of benefitial allele and selection strength.
}
\seealso{
\code{\link[=haplotest]{haplotest()}}
}
