% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frequency_sim.R
\name{Frequency_sim}
\alias{Frequency_sim}
\title{Simulation of haploid haplotype frequencies over time.}
\usage{
Frequency_sim(
  haplotype,
  t = 6,
  tdelta = 10,
  benef_sim,
  rand_start = FALSE,
  Ne = 500,
  sel_dec = FALSE,
  dec_strength = 0.7,
  dec_ne = 0,
  repli = 1
)
}
\arguments{
\item{haplotype}{Binary numeric matrix containing information of haplotype structure. The columns corresponds to the haplotypes and the rows to the SNPs.}

\item{t}{Numeric, number of time points of interest, excluding the initial time point}

\item{rand_start}{Logical. rand_start = TRUE corresponds to random sampling of the starting haplotype frequency from a uniform distribution. If rand_start = FALSE the starting haplotype frequency is set to equal for all haplotypes.}

\item{Ne}{Positive integer, the effective population size.}

\item{sel_dec}{Logical. sel_dec = TRUE corresponds to decreasing selective strength (fitness of selected haplotype) over time}

\item{dec_strength}{Numeric, strength by which selection decreases if sel_dec=TRUE. If sel_dec = FALSE this parameter is ignored.}

\item{dec_ne}{Integer, strength by which Ne decreases over time. dec_ne = number of Ne decreased per generation. Contant Ne if equals to 0.}

\item{repli}{Numeric, specifying the number of replicated populations.}

\item{deltat}{Numeric, number of generations between each pair of time points of interests}

\item{s}{Numeric vector of selection coefficients. s = 0 corresponds to neutral simulations. Can use first element of list from output of benef_sim() as input.}

\item{benef_all}{Numeric vector of position of beneficial alleles. Can use second element of list from output of benef_sim() as input.}
}
\value{
Returns a list. The first item is the haplotype frequency matrix with row being the haplotypes and columns the chosen timepoints. If replicates are present, the column will be the time points of first replicate, followed by the timepoints of second replicates etc. Second item is the fitness of haplotypes.
}
\description{
Simulation of haplotype frequencies over time using a multinomial distribution. Simulations can be run with or without replicated populations, and for constant or changing selection over time. Frequencies are recorded at time points provided in the input data.
}
