% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hap_block.R
\name{hap_block}
\alias{hap_block}
\title{Selection of non-overlapping haplotype block from filtered haplotype block set}
\usage{
hap_block(filt_hap)
}
\arguments{
\item{filt_hap}{Numeric matrix, the filtered haplotype block set, output from MIG()}
}
\value{
Numeric matrix, set of non-overlapping haplotype block location, with the first column being the starting SNP position of the candidate, and second column ending SNP position of the candidate. Each row will corresponds to a single candidate.
}
\description{
The algorithm selects non-overlapping haplotype block from the filtered haplotype block candidate output from the MIG algorithm, by favouring larger haplotype blocks first.
}
