% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hap_err_sim.R
\name{hap_err_sim}
\alias{hap_err_sim}
\title{Simulation of estimated haplotype frequencies with some sample size}
\usage{
hap_err_sim(freq, samsize)
}
\arguments{
\item{freq}{Numeric matrix, with each i,j element being the frequency of haplotype i at sequenced time point j.}

\item{samsize}{Numeric vector or numeric. If numeric vector, a samplesize vector the same length as the ncol(freq). If numeric, the mean sample size, a poisson distribution will be used to determine the samplesize at different timepoints.}
}
\value{
A list, with the first item being the noisy frequency matrix, the second item the samplesize matrix.
}
\description{
Noisy estimated haplotype frequencies are simulated assuming multinomial distribution with some sample size.
}
\seealso{
\code{\link[=haplotest]{haplotest()}}
}
