% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hap_snp_red.R
\name{hap_snp_red}
\alias{hap_snp_red}
\title{Combination of haplotypes}
\usage{
hap_snp_red(red_snp, frequency_matrix)
}
\arguments{
\item{red_snp}{Binary matrix, the reduced haplotype structural matrix with reduced SNPs.}

\item{frequency_matrix}{Numeric matrix, with each i,j element being the frequency of haplotype i at sequenced time point j.}
}
\value{
A reduced haplotype frequency matrix (if haplotypes are combined due to reduced SNPs). Remains unchanged from the input frequency_matrix if no combination of haplotypes occurred.
}
\description{
Combine haplotype frequencies according to the reduced haplotype structural matrix created through SNP based test/haplotype blocks.
}
\seealso{
\code{\link[=haplotest]{haplotest()}}
}
