% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotest.R
\name{haplotest}
\alias{haplotest}
\title{Testing for selection using temporal haplotype/allele frequencies.}
\usage{
haplotest(
  haplotype = NULL,
  haplotype_frequency = NULL,
  allele_frequency = NULL,
  coverage_matrix = NULL,
  deltat = 10,
  Ne = 1000,
  repli = 1,
  test_type = "haplotype",
  test_method = "base",
  p_combine_method = "omnibus",
  freq_est = "known",
  significance = 0.05,
  seed = 2021
)
}
\arguments{
\item{haplotype}{Binary numeric matrix containing information of haplotype structure. The columns corresponds to the haplotypes and the rows to the SNPs.}

\item{haplotype_frequency}{Numeric matrix, with each i,j element being the frequency of haplotype i at sequenced time point j.}

\item{allele_frequency}{Numeric matrix, with each i,j element being the frequency of SNP i at sequenced time point j.}

\item{coverage_matrix}{Numeric matrix, required if frequencies are estimated. Sample size matrix of either haplotype/allele frequencies. If test_method == "hap_filt" for haplotype based test, sample size matrix for both haplotype and allele needs to be present. It should  be in the form of list(haplotype_sample_size, allele_sample_size).}

\item{deltat}{Numeric, number of generations between each pair of time points of interests}

\item{Ne}{Numeric vector with length as number of replicates, containing information of Ne (effective population size) at each replicated population. If Ne changes over time, take as input a numeric matrix, with the column being the replicate position, row being the Ne at each sequenced time points.}

\item{repli}{Numeric, specifying the number of replicated populations.}

\item{test_type}{Factor, can be either "haplotype" or "allele", denoting whether to use haplotype based test or snp based test}

\item{test_method}{Factor, can be "base", "hap_filt" or "hap_block" for haplotype based, denoting basic haplotype based test, haplotype filtering using snp based test, or haplotype block based test. Can only take argument "base" for snp based.}

\item{p_combine_method}{Factor, method of pvalue combination, can be "omnibus" (Futschik, A. et al. 2019), "harmonic" (Wilson, D.J. (2019)), "vovk" (Vovk, V. et al. 2018), "bonferroni", "BH" (Benjamini & Hochberg 1995).}

\item{freq_est}{Factor, whether the frequencies are known or estimated. Input "known" if all frequencies are known, "half" if starting frequencies are known and others estimated, "est" if all are estimated.}

\item{significance}{Numeric, must be between 0 and 1, level of significance for snp based test if "hap_filt" method is used.}

\item{seed}{setting seed of the run}
}
\value{
A numeric vector of p-values from the test after multiple testing corrections.
}
\description{
Selection can be tested using either haplotype or allele frequencies. For haplotype frequencies, 3 methods are available, basic, haplotype filtering using SNPs, haplotype block based. All tests uses the adapted chi-squared test to correct for variance during testing, dependent on whether the frequencies are estimated. For haplotype filtering and haplotype block based method, both haplotype and allele frequencies, as well as the haplotype structure will be required.
}
\references{
Spitzer, K., Pelizzola, M., Futschik, A., (2019), Modifying the Chi-square and the CMH test for population genetic inference: adapting to over-dispersion, The Annals of Applied Statistics 14(1): 202-220.

Futschik, A., Taus, T., Zehetmayer, S., (2019), An omnibus test for the global null hypothesis, Statistical Methods in Medical Research 28(8): 2292-2304.

Wilson, D. J., (2019), The harmonic mean p-value for combining dependent tests, Proceedings of the National Academy of Sciences of the United States of America 116(4): 1195-1200.

Benjamini, Y., Hochberg, Y., (1995), Controlling the false discovery rate: a practical and powerful approach to multiple testing, Journal of the Royal Statistical Society Series B 57(1): 289-300.

Vovk, V., Wang., R., (2018), Combining p-values via averaging, Biometrika 107(4): 791–808.
}
