% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Haplotype_sim.R
\name{Haplotype_sim}
\alias{Haplotype_sim}
\title{Simulate a haplotype structure matrix}
\usage{
Haplotype_sim(nsnp = 500, nhap = 5)
}
\arguments{
\item{nsnp}{Integer, number of SNPs}

\item{nhap}{Integer, number of haplotypes}
}
\value{
Returns a numeric matrix, with all entries being binary, where the rows correspond to the SNPs and the columns to the haplotypes.
}
\description{
Simulate a haplotype structure matrix through random sampling of binary argutments, where the rows correspond to the SNPs and the columns to the haplotypes.
}
