% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merging_function.R
\name{merging_function}
\alias{merging_function}
\title{Vovk based multiple testing correction.}
\usage{
merging_function(pval, r)
}
\arguments{
\item{pval}{Numeric vector of p-values used for combination.}

\item{r}{Numeric, parameter for Vovk's methods.}
}
\value{
Numeric, a single pvalue.
}
\description{
Multiple testing combination method based on Vovk et al. 2018.
}
\references{
Vovk, V., Wang., R., (2018), Combining p-values via averaging, Biometrika 107(4): 791–808.
}
