% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalised_LD.R
\name{normalised_LD}
\alias{normalised_LD}
\title{Computation of normalised linkage disequillibrium.}
\usage{
normalised_LD(SNP_1, SNP_2, haplotype, frequency_matrix, allele_frequency)
}
\arguments{
\item{SNP_1}{Numeric, SNP position of SNP 1}

\item{SNP_2}{Numeric, SNP position of SNP 2}

\item{haplotype}{Binary numeric matrix containing information of haplotype structure. The columns corresponds to the haplotypes and the rows to the SNPs.}

\item{frequency_matrix}{Numeric matrix, with each i,j element being the frequency of haplotype i at sequenced time point j.}

\item{allele_frequency}{Numeric matrix, with each i,j element being the frequency of SNP i at sequenced time point j.}
}
\value{
A numeric vector with its first element being the normalised LD, second element the estimated variance of normalised LD.
}
\description{
Compute Linkage Disequillibrium(LD) between 2 SNPs, return normalised LD (Lewontin, R. C. 1964) and estimated variance of normalised LD (Zapata, C. et al. 1997).
}
\references{
Lewontin, R. C., (1964), The Interaction of Selection and Linkage. I. General Considerations; Heterotic Models, Genetics 49(1): 49–67.

Zapata, C., Alvarez, G., Carollo, C., (1997), Approximate variance of the standardized measure of gametic disequilibrium D’, American journal of human genetics 61(3): 771–774.
}
